lots of features and bug fixes.

+++ src/main/kotlin/com/github/bpm/objects/quantum/QuantumScreen.kt
@@ -1,6 +1,6 @@
 package com.github.bpm.objects.quantum
 
-import com.github.bpm.render.EditorRenderer
+import com.github.bpm.render.GraphRenderer
 import com.github.bpm.render.PropertiesRenderer
 import com.github.bpm.util.Gui
 import com.mojang.blaze3d.vertex.PoseStack
@@ -10,11 +10,17 @@ import imgui.extension.nodeditor.NodeEditorContext
 import net.minecraft.client.gui.screens.Screen
 import net.minecraft.network.chat.TextComponent
 
-class QuantumScreen(private val tile: QuantumTile) : Screen(TextComponent("Quantum Screen")) {
+class QuantumScreen(val tile: QuantumTile) : Screen(TextComponent("Quantum Screen")) {
     private val context = NodeEditorContext()
     private val nodeContext = ImNodesContext()
-    private val editorRenderer = EditorRenderer(tile.graph)
     private val propertiesRenderer = PropertiesRenderer(tile.graph)
+    private val graphRenderer = GraphRenderer(tile.graph)
+    private var closed = false
+
+    override fun init() {
+        tile.requestGraph()
+        minecraft?.mouseHandler?.releaseMouse()
+    }
 
     override fun render(stack: PoseStack, var1: Int, var2: Int, var3: Float) = Gui.frame {
         Gui.dockspace("Quantum Screen", propertiesRenderer::render, ::renderViewport)
@@ -22,15 +28,22 @@ class QuantumScreen(private val tile: QuantumTile) : Screen(TextComponent("Quant
 
     fun renderViewport() {
         ImNodes.editorContextSet(nodeContext)
-        Gui.nodeGraph("Quantum Screen", editorRenderer::render)
-        editorRenderer.poll()
+        Gui.nodeGraph("Quantum Screen", graphRenderer::render)
+        graphRenderer.poll()
     }
 
+    fun close() {
+        if (!closed) {
+            graphRenderer.beforeClose()
+            tile.pushGraph()
+            context.destroy()
+            nodeContext.destroy()
+            closed = true
+        }
+    }
 
     override fun onClose() {
-        tile.pushGraph()
-        context.destroy()
-        nodeContext.destroy()
+        close()
         super.onClose()
     }
 }
